refactor: update policy cache logic
